refactor(ui): useGetAvatarChipLeftComponentAndVariant
