+++ packages/twenty-ui/src/display/chip/avatar/types/avatarChipsVariant.type.ts
@@ -1,0 +1,4 @@
+export enum AvatarChipVariant {
+  Regular = 'regular',
+  Transparent = 'transparent',
+}
