+++ packages/twenty-ui/src/display/index.ts
@@ -7,11 +7,14 @@ export * from './avatar/types/AvatarType';
 export * from './banner/components/Banner';
 export * from './checkmark/components/AnimatedCheckmark';
 export * from './checkmark/components/Checkmark';
-export * from './chip/components/AvatarChip';
+export * from './chip/avatar/components/AvatarChip';
+export * from './chip/avatar/components/AvatarChipLeftComponent';
+export * from './chip/avatar/components/LinkAvatarChip';
+export * from './chip/avatar/types/avatarChipsCommonProps.type';
+export * from './chip/avatar/types/avatarChipsVariant.type';
+export * from './chip/avatar/utils/getAvatarChipsVariant.util';
 export * from './chip/components/Chip';
-export * from './chip/components/LinkAvatarChip';
 export * from './chip/components/LinkChip';
-export * from './chip/hooks/useGetAvatarChipLeftComponentAndVariant';
 export * from './color/components/ColorSample';
 export * from './icon/components/IconAddressBook';
 export * from './icon/components/IconGmail';
